/age: Fix floating point operations
Apparently in java it doesn't cast to a floating point value until after
the computation, so e.g. 100 / 60 = 1.0, and 100 / 60.0 = 1.6...

+++ src/AgeCommand.java
@@ -143,10 +143,10 @@ public class AgeCommand implements CommandExecutor {
 						 || player.canSee(on_player))) {
 					int played_ticks = on_player.getStatistic(Statistic.PLAY_ONE_MINUTE);
 					int played_minutes = played_ticks / (20 * 60);
-					double played_hours = played_minutes / 60;
-					double played_days = played_hours / 24;
-					double played_years = played_days / 365;
-					double avg_hours = played_hours / (days + 1);
+					double played_hours = played_minutes / 60.0;
+					double played_days = played_hours / 24.0;
+					double played_years = played_days / 365.0;
+					double avg_hours = played_hours / (double) (days + 1);
 					msg += String.format(" They have played for %d minutes = %.1f hours = %.1f days = %.4f years. For an average of %.2f hours per day.",
 							played_minutes, played_hours, played_days, played_years, avg_hours);
 				} else {
